@@ -1,33 +1,35 @@
 import { Modem, InitializationError, ConnectionMonitor } from "../lib/index";
-import * as fs from "fs";
-import * as path from "path";
 //@ts-ignore: we may un comment
 import * as repl from "repl";
 
 import "colors";
 
 import * as _debug from "debug";
-let debug= _debug("main");
-debug.enabled= true;
+let debug = _debug("main");
+debug.enabled = true;
 
-(async ()=>{
+process.on("unhandledRejection", error=> {
+    throw error;
+});
+
+(async () => {
 
     debug("Started, looking for connected modem...");
 
-    let accessPoint= await ConnectionMonitor.getInstance().evtModemConnect.waitFor();
+    let accessPoint = await ConnectionMonitor.getInstance().evtModemConnect.waitFor();
 
     let modem: Modem;
 
-    try{
+    try {
 
-        modem= await Modem.create({
+        modem = await Modem.create({
             "dataIfPath": accessPoint.dataIfPath,
             "unlock": { "pinFirstTry": "0000", "pinSecondTry": "1234" }
         });
 
-    }catch(error){
+    } catch (error) {
 
-        let initializationError= error as InitializationError;
+        let initializationError = error as InitializationError;
 
         debug(initializationError);
 
@@ -35,7 +37,7 @@ debug.enabled= true;
 
     }
 
-    modem.evtTerminate.attachOnce(error=>{
+    modem.evtTerminate.attachOnce(error => {
 
         debug("Modem terminate", { error });
 
@@ -43,23 +45,56 @@ debug.enabled= true;
 
     });
 
-    let messageText = fs.readFileSync(path.join(__dirname, "messageText.txt").replace(/dist/, "src"), "utf8");
     //let messageText= "foo bar";
 
-    let joseph= "0636786385";
+    // cSpell:disable
+    const messageText = [
+        `Un mal qui répand la terreur,`,
+        `Mal que le Ciel en sa fureur`,
+        `Inventa pour punir les crimes de la terre,`,
+        `La Peste (puisqu’il faut l’appeler par son nom),`,
+        `Capable d’enrichir en un jour l’Achéron,`,
+        `Faisait aux Animaux la guerre.`,
+        `Ils ne mouraient pas tous, mais tous étaient frappés :`,
+        `On n’en voyait point d’occupés`,
+        `À chercher le soutien d’une mourante vie ;`,
+        `Nul mets n’excitait leur envie ;`,
+        `Ni Loups ni Renards n’épiaient`,
+        `La douce et l’innocente proie ;`,
+        `Les Tourterelles se fuyaient :`,
+        `Plus d’amour, partant plus de joie.`,
+        `Le Lion tint conseil, et dit : « Mes chers amis,`,
+        `Je crois que le Ciel a permis`,
+        `Pour nos péchés cette infortune.`,
+        `Que le plus coupable de nous`,
+        `Se sacrifie aux traits du céleste courroux ;`,
+        `Peut-être il obtiendra la guérison commune.`,
+        `L’histoire nous apprend qu’en de tels accidents`,
+        `On fait de pareils dévouements.`,
+        `Ne nous flattons donc point, voyons sans indulgence`,
+        `L’état de notre conscience.`,
+        `Pour moi, satisfaisant mes appétits gloutons,`,
+        `J’ai dévoré force moutons.`,
+        `Que m’avaient-ils fait ? nulle offense ;`,
+        `Même il m’est arrivé quelquefois de manger`,
+        `Le berger.`
+    ].join("\n");
+    /* spell-checker: enable */
 
-    let sentMessageId= await modem.sendMessage(joseph, messageText);
+    let joseph = "0636786385";
+
+    let sentMessageId = await modem.sendMessage(joseph, messageText);
 
     debug("SendDate( used as id): ", sentMessageId);
 
-    await new Promise(resolve=> setTimeout(resolve, 60000));
+    await new Promise(resolve => setTimeout(resolve, 60000));
 
     debug("Manual termination of the modem");
 
     modem.terminate();
 
     console.assert(modem.isTerminated === true);
-    
+
     /*
 
     let { context } = repl.start({
